Drop Ledger module name
