--- lib/ledger_sync/ledgers/xero/config.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-LedgerSync.register_ledger(:xero) do |config|
-  config.name = 'Xero'
-end
