--- lib/ledger_sync/ledgers/xero/contact/serializer.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module LedgerSync
-  module Ledgers
-    module Xero
-      class Contact
-        class Serializer < Xero::Serializer
-          attribute 'ContactID',
-                    resource_attribute: :ledger_id
-          attribute :Name
-          attribute :EmailAddress
-        end
-      end
-    end
-  end
-end
