--- lib/ledger_sync/ledgers/xero/deserializer.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module LedgerSync
-  module Ledgers
-    module Xero
-      class Deserializer < LedgerSync::Deserializer
-      end
-    end
-  end
-end
