--- lib/ledger_sync/ledgers/xero/resource.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module LedgerSync
-  module Ledgers
-    module Xero
-      class Resource < LedgerSync::Resource
-      end
-    end
-  end
-end
