--- lib/ledger_sync/ledgers/xero/serializer.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module LedgerSync
-  module Ledgers
-    module Xero
-      class Serializer < LedgerSync::Serializer
-      end
-    end
-  end
-end
