--- lib/ledger_sync/ledgers/xero/version.rb
@@ -1,7 +1,0 @@
-module LedgerSync
-  module Ledgers
-    module Xero
-      VERSION = "0.1.0"
-    end
-  end
-end
