+++ lib/ledger_sync/xero.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 require 'ledger_sync'
+require_relative 'xero/config'
 
 module LedgerSync
   module Xero
-    class Error < LedgerSync::Error; end
   end
 end
