+++ lib/ledger_sync/xero/config.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+LedgerSync.register_ledger(:xero, { root_path: 'ledger_sync/xero' }) do |config|
+  config.name = 'Xero'
+end
