+++ lib/ledger_sync/xero/contact/deserializer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Xero
+    class Contact
+      class Deserializer < Xero::Deserializer
+        attribute :ledger_id,
+                  hash_attribute: 'ContactID'
+        attribute :Name
+        attribute :EmailAddress
+      end
+    end
+  end
+end
