+++ lib/ledger_sync/xero/contact/operations/create.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Xero
+    class Contact
+      module Operations
+        class Create < Xero::Operation::Create
+          class Contract < LedgerSync::Ledgers::Contract
+            params do
+              required(:external_id).filled(:string)
+              required(:ledger_id).value(:nil)
+              required(:Name).maybe(:string)
+              required(:EmailAddress).maybe(:string)
+            end
+          end
+        end
+      end
+    end
+  end
+end
