+++ lib/ledger_sync/xero/deserializer.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Xero
+    class Deserializer < LedgerSync::Deserializer
+    end
+  end
+end
