+++ lib/ledger_sync/xero/resource.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Xero
+    class Resource < LedgerSync::Resource
+    end
+  end
+end
