+++ lib/ledger_sync/xero/serializer.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module LedgerSync
+  module Xero
+    class Serializer < LedgerSync::Serializer
+    end
+  end
+end
