+++ lib/ledger_sync/xero/version.rb
@@ -1,0 +1,5 @@
+module LedgerSync
+  module Xero
+    VERSION = "0.1.0"
+  end
+end
