+++ spec/ledger_sync/ledgers/xero/resources/version_spec.rb
@@ -1,5 +1,5 @@
-RSpec.describe LedgerSync::Ledgers::Xero do
+RSpec.describe LedgerSync::Xero do
   it 'has a version number' do
-    expect(LedgerSync::Ledgers::Xero::VERSION).not_to be nil
+    expect(LedgerSync::Xero::VERSION).not_to be nil
   end
 end
